Module parity_plonk::commitment_scheme

Ideally we should cleanly abstract away the polynomial commitment scheme We note that PLONK makes use of the linearisation technique conceived in SONIC [Mary Maller].

This technique implicitly requires the commitment scheme to be homomorphic. Merkle Tree like techniques such as FRI are not homomorphic and therefore for PLONK to be usable with all commitment schemes without modification, one would need to remove the lineariser

Structs

CommitKey

CommitKey is used to commit to a polynomial which is bounded by the max_degree.

OpeningKey

Opening Key is used to verify opening proofs made about a committed polynomial.

PublicParameters

The Public Parameters can also be referred to as the Structured Reference String (SRS). It is available to both the prover and verifier and allows the verifier to efficiently verify and make claims about polynomials up to and including a configured degree.