Struct parity_plonk::prelude::VerifierKey

pub struct VerifierKey { /* fields omitted */ }

PLONK circuit Verification Key.

This structure is used by the Verifier in order to verify a Proof.

Implementations

impl VerifierKey

pub const fn padded_gates(&self) -> u32

Returns the Circuit size padded to the next power of two.