Module parity_plonk::proof_system

Proving system

Structs

Proof

A Proof is a composition of Commitments to the Witness, Permutation, Quotient, Shifted and Opening polynomials as well as the ProofEvaluations.

Prover

Abstraction structure designed to construct a circuit and generate Proofs for it.

ProverKey

PLONK circuit Proving Key.

Verifier

Abstraction structure designed verify Proofs.

VerifierKey

PLONK circuit Verification Key.