Struct parity_plonk::proof_system::Prover

pub struct Prover {
    pub prover_key: Option<ProverKey>,
    pub preprocessed_transcript: Transcript,
    // some fields omitted
}

Abstraction structure designed to construct a circuit and generate Proofs for it.

Fields
prover_key: Option<ProverKey>

ProverKey which is used to create proofs about a specific PLONK circuit

preprocessed_transcript: Transcript

Store the messages exchanged during the preprocessing stage This is copied each time, we make a proof

Implementations

impl Prover

pub fn composer_mut(&mut self) -> &mut TurboComposer

Mutable borrow of TurboComposer.

pub fn preprocess(&mut self, commit_key: &CommitKey) -> Result<(), Error>

Preprocesses the underlying constraint system.

impl Prover

pub fn new(label: &'static [u8]) -> Prover

Creates a new Prover instance.

pub fn with_size(label: &'static [u8], size: usize) -> Prover

Creates a new Prover object with some expected size.

pub const fn gates(&self) -> u32

Returns the number of gates in the circuit thet the Prover actually stores inside.

pub fn clear_witness(&mut self)

Resets the witnesses in the prover object. This function is used when the user wants to make multiple proofs with the same circuit.

pub fn clear(&mut self)

Clears all data in the Prover instance. This function is used when the user wants to use the same Prover to make a Proof regarding a different circuit.

pub fn key_transcript(&mut self, label: &'static [u8], message: &[u8])

Keys the Transcript with additional seed information Wrapper around Transcript::append_message.

pub fn prove_with_preprocessed(
&self,
commit_key: &CommitKey,
prover_key: &ProverKey
) -> Result<Proof, Error>

Creates a [Proof] that demonstrates that a circuit is satisfied.

Note

If you intend to construct multiple Proofs with different witnesses, after calling this method, the user should then call Prover::clear_witness. This is automatically done when Prover::prove is called.

pub fn prove(&mut self, commit_key: &CommitKey) -> Result<Proof, Error>

Proves a circuit is satisfied, then clears the witness variables If the circuit is not pre-processed, then the preprocessed circuit will also be computed.